changed double to decimal

+++ src/main/resources/init.sql
@@ -27,7 +27,7 @@ DROP TABLE IF EXISTS `Files`;
 CREATE TABLE `Files` (
   `id` int(11) NOT NULL AUTO_INCREMENT,
   `path` varchar(210) NOT NULL,
-  `size` double NOT NULL,
+  `size` DECIMAL(10,2) NOT NULL,
   `uploadDate` datetime DEFAULT NULL,
   `filename` varchar(70) DEFAULT NULL,
   `extension` varchar(45) DEFAULT NULL,
